feature(app): Add user management
This commit introduces users to the bot. Each user gets added
to the `users` table when they start using the bot, including a
Telegram ID and their name.

As part of this, a `/start` command is included which handles the
user registration flow.

Test coverage is a bit slim on this, which will be revisited at
a later date.